==> 2_Categorizar_negocios_metadata.ipynb <==
import pandas as pd
import nltk
from nltk.corpus import stopwords
nltk.download('punkt')

En este archivo encontrarás el código para poder saber cuáles son las palabras más utilizadas en la columna category, esto para poder asignar 1 solo elemento que describa el negocio, ya que la mayoría de las filas contienen más de una descripción de la categoría del negocio. 

metadata=pd.read_csv('df_metadata_limpia.csv')
metadata

non_string_values = metadata['category'].apply(lambda x: type(x) != str)
non_string_values.unique()

metadata['category'] = metadata['category'].fillna('')

metadata['category'] = metadata['category'].astype(str)

Quitamos las stopwords

import nltk
nltk.download('stopwords')
stopwords = nltk.corpus.stopwords.words(['spanish', 'english'])

# Función para tokenizar y filtrar stopwords
def tokenize_and_filter(text):
    words = text.split()  # Divide la cadena en palabras
    filtered_words = [word for word in words if word.lower() not in stopwords]  # Filtra las stopwords
    return filtered_words


import re 

metadata['category'] = metadata['category'].apply(lambda x: re.sub(r'[^\w\s]', '', x))


metadata['filtrado'] = metadata['category'].apply(tokenize_and_filter)
metadata

top_N=300

words=metadata.filtrado.tolist()
words=[word for list_ in words for word in list_]
words

word_dist= nltk.FreqDist(words)
rslt=pd.DataFrame(word_dist.most_common(top_N),columns=['Word','Frecuencia'])
rslt

categorias_palabras_clave = {
    'Restaurante': ['restaurant','pizza','bar','food','takeout','coffee','delivery','venue','sandwich','cafe','mexican','breakfast','diner'],
    'Agencia': ['agency'],
    'Fotógrafo': ['photographer'],
    'Auto': ['transportation','auto','repair','car','truck','tire','motorcycle','trailer','mechanic','chevrolet'],
    'Belleza_Spa':['salon','beauty','hair','nail','spa','barber','waxing','massage'],
    'Mascotas': ['pet','veterinarian','dog','animal'],
    'Salud': ['dentist','doctor','medical','health','dental','surgeon','pharmacy','drug','physician','hospital','optometrist','pathologist'],
    'Iglesia':['church','baptist','wedding','religious','catholic'],
    'Casa':['home','contractor','furniture'],
    'Oficina':['office','corporate'],
    'Escuela':['school'],
    'Banco':['bank','atm','ATM','money'],
    'Deportes':['sports','club','physical','gym','GYM','body','fitness','ski'],
    'Atraccion Turistica':['tourist', 'attraction','park'],
    'Tienda':['liquor','cell','phone','electronic','convenience'],
    'Sin Categoria':[''],
    'Hospedaje':['lodging','hotel'],
    'Fabrica':['fabric']

}

def clasificar_categorias(categorias):
    for categoria, palabras_clave in categorias_palabras_clave.items():
        for palabra_clave in palabras_clave:
            if palabra_clave in categorias:
                return categoria
    return 'Otros'  # Si no se encuentra ninguna categoría, se clasifica como 'otros'


# Aplicar la función a la columna 'categorias' y crear una nueva columna 'clasificacion'
metadata['clasificacion'] = metadata['filtrado'].apply(clasificar_categorias)

metadata[metadata['clasificacion']=='Otros']

metadata.to_csv('metadata2.csv', index=False )
