==> 2_limpieza_california.ipynb <==
import pandas as pd
import numpy as np
import json
import datetime
import re #Expresiones regulares
import emoji 
import demoji 

Función para quitar emojis

def remove_emojis(text):
    return demoji.replace(text, '')

def remove_emojis_2(text):
    if type(text) == str:
        return demoji.replace(text, '')
    else:
        return None 

# California

Cargamos los datos

df1_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/1.json', lines=True)
df2_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/2.json', lines=True)
df3_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/3.json', lines=True)
df4_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/4.json', lines=True)
df5_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/5.json', lines=True)
df6_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/6.json', lines=True)
df7_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/7.json', lines=True)
df8_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/8.json', lines=True)
df9_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/9.json', lines=True)
df10_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/10.json', lines=True)
df11_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/11.json', lines=True)
df12_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/12.json', lines=True)
df13_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/13.json', lines=True)
df14_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/14.json', lines=True)
df15_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/15.json', lines=True)
df16_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/16.json', lines=True)
df17_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/17.json', lines=True)
df18_ny = pd.read_json('../Archivos_originales/reviews-estados/review-California/18.json', lines=True)


dataframes_ny = [df1_ny, df2_ny, df3_ny, df4_ny, df5_ny, 
                         df6_ny, df7_ny, df8_ny, df9_ny, df10_ny, 
                         df11_ny,df12_ny,df13_ny,df14_ny,df15_ny,df16_ny,df17_ny,df18_ny]


# df_ny = crear_df('New_York',19)
# df_ny
# NÚMERO DE DATOS = 2700000

Se checan el número de nulos que contiene el dataframe

for i, df in enumerate(dataframes_ny):
    nulos = df.isnull().sum()
    print(f"DataFrame {i+1} - Missing Values:\n{nulos}\n")

Seleccion de columnas 

columnas= ['name','time','rating','text','gmap_id']
df1_ny = df1_ny[columnas]
df2_ny = df2_ny[columnas]
df3_ny =df3_ny[columnas]
df4_ny = df4_ny[columnas]
df5_ny =df5_ny[columnas]
df6_ny = df6_ny[columnas]
df7_ny = df7_ny[columnas]
df8_ny = df8_ny[columnas]
df9_ny = df9_ny[columnas]
df10_ny = df10_ny[columnas]
df11_ny = df11_ny[columnas]
df12_ny =df12_ny[columnas]
df13_ny =df13_ny[columnas]
df14_ny =df14_ny[columnas]
df15_ny =df15_ny[columnas]
df16_ny =df16_ny[columnas]
df17_ny =df17_ny[columnas]
df18_ny =df18_ny[columnas]



Se eliminan los nulos que contiene la columna text

df1_ny = df1_ny.dropna(subset=['text'])
df2_ny = df2_ny.dropna(subset=['text'])
df3_ny = df3_ny.dropna(subset=['text'])
df4_ny = df4_ny.dropna(subset=['text'])
df5_ny = df5_ny.dropna(subset=['text'])
df6_ny = df6_ny.dropna(subset=['text'])
df7_ny = df7_ny.dropna(subset=['text'])
df8_ny = df8_ny.dropna(subset=['text'])
df9_ny = df9_ny.dropna(subset=['text'])
df10_ny = df10_ny.dropna(subset=['text'])
df11_ny = df11_ny.dropna(subset=['text'])
df12_ny = df12_ny.dropna(subset=['text'])
df13_ny = df13_ny.dropna(subset=['text'])
df14_ny = df14_ny.dropna(subset=['text'])
df15_ny = df15_ny.dropna(subset=['text'])
df16_ny = df16_ny.dropna(subset=['text'])
df17_ny = df17_ny.dropna(subset=['text'])
df18_ny = df18_ny.dropna(subset=['text'])



Convertimos la columna time int a formato fecha

df1_ny['time'] = df1_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df2_ny['time'] = df2_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df3_ny['time'] = df3_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df4_ny['time'] = df4_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df5_ny['time'] = df5_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df6_ny['time'] = df6_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df7_ny['time'] = df7_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df8_ny['time'] = df8_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df9_ny['time'] = df9_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df10_ny['time'] = df10_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df11_ny['time'] = df11_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df12_ny['time'] = df12_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df13_ny['time'] = df13_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df14_ny['time'] = df14_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df15_ny['time'] = df15_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df16_ny['time'] = df16_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df17_ny['time'] = df17_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))
df18_ny['time'] = df18_ny['time'].apply(lambda x: pd.to_datetime(x, unit='ms'))



Se crean las columnas anio, mes y día

df1_ny['anio'] = df1_ny['time'].dt.year
df2_ny['anio'] = df2_ny['time'].dt.year
df3_ny['anio'] = df3_ny['time'].dt.year
df4_ny['anio'] = df4_ny['time'].dt.year
df5_ny['anio'] = df5_ny['time'].dt.year
df6_ny['anio'] = df6_ny['time'].dt.year
df7_ny['anio'] = df7_ny['time'].dt.year
df8_ny['anio'] = df8_ny['time'].dt.year
df9_ny['anio'] = df9_ny['time'].dt.year
df10_ny['anio'] = df10_ny['time'].dt.year
df11_ny['anio'] = df11_ny['time'].dt.year
df12_ny['anio'] = df12_ny['time'].dt.year
df13_ny['anio'] = df13_ny['time'].dt.year
df14_ny['anio'] = df14_ny['time'].dt.year
df15_ny['anio'] = df15_ny['time'].dt.year
df16_ny['anio'] = df16_ny['time'].dt.year
df17_ny['anio'] = df17_ny['time'].dt.year
df18_ny['anio'] = df18_ny['time'].dt.year



df1_ny['mes'] = df1_ny['time'].dt.month
df2_ny['mes'] = df2_ny['time'].dt.month
df3_ny['mes'] = df3_ny['time'].dt.month
df4_ny['mes'] = df4_ny['time'].dt.month
df5_ny['mes'] = df5_ny['time'].dt.month
df6_ny['mes'] = df6_ny['time'].dt.month
df7_ny['mes'] = df7_ny['time'].dt.month
df8_ny['mes'] = df8_ny['time'].dt.month
df9_ny['mes'] = df9_ny['time'].dt.month
df10_ny['mes'] = df10_ny['time'].dt.month
df11_ny['mes'] = df11_ny['time'].dt.month
df12_ny['mes'] = df12_ny['time'].dt.month
df13_ny['mes'] = df13_ny['time'].dt.month
df14_ny['mes'] = df14_ny['time'].dt.month
df15_ny['mes'] = df15_ny['time'].dt.month
df16_ny['mes'] = df16_ny['time'].dt.month
df17_ny['mes'] = df17_ny['time'].dt.month
df18_ny['mes'] = df18_ny['time'].dt.month




df1_ny['dia'] = df1_ny['time'].dt.day
df2_ny['dia'] = df2_ny['time'].dt.day
df3_ny['dia'] = df3_ny['time'].dt.day
df4_ny['dia'] = df4_ny['time'].dt.day
df5_ny['dia'] = df5_ny['time'].dt.day
df6_ny['dia'] = df6_ny['time'].dt.day
df7_ny['dia'] = df7_ny['time'].dt.day
df8_ny['dia'] = df8_ny['time'].dt.day
df9_ny['dia'] = df9_ny['time'].dt.day
df10_ny['dia'] = df10_ny['time'].dt.day
df11_ny['dia'] = df11_ny['time'].dt.day
df12_ny['dia'] = df12_ny['time'].dt.day
df13_ny['dia'] = df13_ny['time'].dt.day
df14_ny['dia'] = df14_ny['time'].dt.day
df15_ny['dia'] = df15_ny['time'].dt.day
df16_ny['dia'] = df16_ny['time'].dt.day
df17_ny['dia'] = df17_ny['time'].dt.day
df18_ny['dia'] = df18_ny['time'].dt.day


df1_ny = df1_ny.drop(['time'],axis=1)
df2_ny = df2_ny.drop(['time'],axis=1)
df3_ny= df3_ny.drop(['time'],axis=1)
df4_ny = df4_ny.drop(['time'],axis=1)
df5_ny = df5_ny.drop(['time'],axis=1)
df6_ny = df6_ny.drop(['time'],axis=1)
df7_ny = df7_ny.drop(['time'],axis=1)
df8_ny = df8_ny.drop(['time'],axis=1)
df9_ny = df9_ny.drop(['time'],axis=1)
df10_ny = df10_ny.drop(['time'],axis=1)
df11_ny = df11_ny.drop(['time'],axis=1)
df12_ny= df12_ny.drop(['time'],axis=1)
df13_ny = df13_ny.drop(['time'],axis=1)
df14_ny= df14_ny.drop(['time'],axis=1)
df15_ny = df15_ny.drop(['time'],axis=1)
df16_ny = df16_ny.drop(['time'],axis=1)
df17_ny = df17_ny.drop(['time'],axis=1)
df18_ny = df18_ny.drop(['time'],axis=1)


df1_ny

Se seleccionan los años 2017 a 2021

df1_ny=df1_ny[df1_ny['anio']>=2017]
df2_ny=df2_ny[df2_ny['anio']>=2017]
df3_ny=df3_ny[df3_ny['anio']>=2017]
df4_ny=df4_ny[df4_ny['anio']>=2017]
df5_ny=df5_ny[df5_ny['anio']>=2017]
df6_ny=df6_ny[df6_ny['anio']>=2017]
df7_n=df7_ny[df7_ny['anio']>=2017]
df8_ny=df8_ny[df8_ny['anio']>=2017]
df9_ny=df9_ny[df9_ny['anio']>=2017]
df10_ny=df10_ny[df10_ny['anio']>=2017]
df11_ny=df11_ny[df11_ny['anio']>=2017]
df12_ny=df12_ny[df12_ny['anio']>=2017]
df13_ny=df13_ny[df13_ny['anio']>=2017]
df14_ny=df14_ny[df14_ny['anio']>=2017]
df15_ny=df15_ny[df15_ny['anio']>=2017]
df16_ny=df16_ny[df16_ny['anio']>=2017]
df17_ny=df17_ny[df17_ny['anio']>=2017]
df18_ny=df18_ny[df18_ny['anio']>=2017]



Se concatenan todos los dataframes 

dataframes_ny = [df1_ny, df2_ny, df3_ny, df4_ny, df5_ny, 
                         df6_ny, df7_ny, df8_ny, df9_ny, df10_ny, 
                         df11_ny,df12_ny,df13_ny,df14_ny,df15_ny,df16_ny,df17_ny,df18_ny]
df_ny = pd.concat(dataframes_ny, ignore_index=True)
df_ny

Se verifica si existen duplicados 

df_ny.duplicated().sum()


Muestra los datos duplicados 

df_ny[df_ny.duplicated(keep='first')]

df_ny.drop_duplicates(inplace=True)

df_ny.duplicated().sum()

df_ny

Todas a minúsculas y se quitan emojis para columnas name y text

df_ny['name']= df_ny['name'].apply(remove_emojis)
df_ny['text']= df_ny['text'].apply(remove_emojis)
df_ny['name']=df_ny['name'].apply(lambda x: x.lower() if x is not None else '')
df_ny['text']=df_ny['text'].apply(lambda x: x.lower() if x is not None else '')
df_ny

+ La columna text tiene este tipo de formato --> '(translated by google) very good store\n\n(original)\nmuy buena tienda.'
+ Se transforma para sólo dejar el texto en inglés. 

df_ny['text'][1415087]

df_ny['text']= df_ny['text'].apply(lambda x: re.sub(r'\([^)]*\)|[\u4e00-\u9fff]', '', x) )
df_ny['text']= df_ny['text'].apply(lambda x: x.split('\n\n\n')[0].strip())
df_ny

df_ny['text'] = df_ny['text'].apply(lambda x: re.sub(r'[^\w\s]', '', x))
df_ny['name'] = df_ny['name'].apply(lambda x: re.sub(r'[^\w\s]', '', x))



df_ny['text'][1415087]

ny_anios =pd.DataFrame(df_ny.anio.value_counts())
ny_anios= ny_anios.reset_index()
ny_anios= ny_anios.rename(columns={"index": "Anio", "anio": "NY"})
ny_anios

Se renombran las columnas 

nombres_nuevos= {'name':'Usuario', 'rating':'Rating_Usuario', 'text':'Review', 'gmap_id':'gmap_id', 'anio':'Anio', 'mes':'Mes', 'name_y':'Empresa',
       'address':'Direccion', 'avg_rating':'AVG_rating_empresa', 'category':'Categoria_empresa', 'num_of_reviews':'Num_reviews', 'hours':'Horario'}
df_ny.rename(columns=nombres_nuevos, inplace=True)

df_ny

df_ny['Estado']='California'


df_ny

df_ny= df_ny[df_ny['Anio']>=2017]

df_ny= df_ny.reset_index()

df_ny.columns

df_ny=df_ny.drop(['level_0'], axis=1)

df_ny=df_ny.drop(['index'], axis=1)

df_ny

Exportar a csv el archivo limpio

df_ny.to_csv('df_california_limpio.csv',index=False)
