==> 2_limpieza_metadata.ipynb <==
import pandas as pd
import numpy as np
import json
import datetime
import re #Expresiones regulares
import emoji 
import demoji 

Función para quitar emojis

def remove_emojis(text):
    return demoji.replace(text, '')

def remove_emojis_2(text):
    if type(text) == str:
        return demoji.replace(text, '')
    else:
        return None 

ESTADOS: Función para unir todos los archivos en un solo dataframe

def crear_df(estado,num_archivos):
    df_1 = pd.DataFrame()
    for i in range(1, num_archivos):
        df = pd.read_json(f'../Google Maps/reviews-estados/review-{estado}/{i}.json', lines=True)
        df_1 = pd.concat([df_1, df], ignore_index=True)

    return df

METADATA: Función para unir todos los archivos en un solo dataframe

def crear_df_metadata(num_archivos):
    df_1 = pd.DataFrame()
    for i in range(1, num_archivos):
        df = pd.read_json(f'../Google Maps/metadata-sitios/{i}.json', lines=True)
        df_1 = pd.concat([df_1, df], ignore_index=True)

    return df

df1 = pd.read_json('../Archivos_originales/metadata-sitios/1.json', lines=True)
df2= pd.read_json('../Archivos_originales/metadata-sitios/2.json', lines=True)
df3 = pd.read_json('../Archivos_originales/metadata-sitios/3.json', lines=True)
df4 = pd.read_json('../Archivos_originales/metadata-sitios/4.json', lines=True)
df5 = pd.read_json('../Archivos_originales/metadata-sitios/5.json', lines=True)
df6 = pd.read_json('../Archivos_originales/metadata-sitios/6.json', lines=True)
df7 = pd.read_json('../Archivos_originales/metadata-sitios/7.json', lines=True)
df8 = pd.read_json('../Archivos_originales/metadata-sitios/8.json', lines=True)
df9 = pd.read_json('../Archivos_originales/metadata-sitios/9.json', lines=True)
df10 = pd.read_json('../Archivos_originales/metadata-sitios/10.json', lines=True)
df11= pd.read_json('../Archivos_originales/metadata-sitios/11.json', lines=True)


dataframes_ny = [df1, df2, df3, df4, df5, 
                         df6, df7, df8, df9, df10, 
                         df11]

for i, df in enumerate(dataframes_ny):
    nulos = df.isnull().sum()
    print(f"DataFrame {i+1} - Missing Values:\n{nulos}\n")

metadata = pd.concat(dataframes_ny, ignore_index=True)

# Metadata

Se verifican los nulos 

metadata.isnull().sum()

Selección de columnas con las que se va a trabajar

metadata=metadata[['name','gmap_id','address','avg_rating','category','hours','num_of_reviews']]

Se eliminan los nulos para la columna name

metadata = metadata.dropna(subset=['name'])

metadata.isnull().sum()

Columna Category.Para cada dato que es lista se convierte a texto. 

metadata['category'] = metadata['category'].apply(lambda x: ', '.join(x) if x is not None else '')
metadata

VERIFICACIÓN DE DUPLICADOS:  se tuvo que quitar la columna 'hours' para poder hacer la verificación 

metadata_2 = metadata[['name', 'gmap_id', 'address', 'avg_rating', 'category',
       'num_of_reviews']]
metadata_2

metadata_2.duplicated().sum()

Visualizar los registros que están duplicados

metadata_2[metadata_2.duplicated(keep='first')]

metadata_2[metadata_2['name']=='Advanced Home Medical Equipment']

Eliminamos los duplicados

metadata_2.drop_duplicates(inplace=True)
metadata_2

metadata_2.duplicated().sum()

metadata_2[metadata_2['name']=='Advanced Home Medical Equipment']

resultado = metadata_2.merge(metadata['hours'], left_index=True, right_index=True)
resultado

resultado.isna().sum()

Se quitan los emojis y todo a minúsculas para las columnas name,category y address

resultado['name']= resultado['name'].apply(remove_emojis)
resultado['category']=resultado['category'].apply(remove_emojis)
resultado['address']=resultado['address'].apply(remove_emojis_2)
resultado['name']=resultado['name'].apply(lambda x: x.lower())
resultado['category']=resultado['category'].apply(lambda x: x.lower() if x is not None else '')
resultado['address']=resultado['address'].apply(lambda x: x.lower() if x is not None else '')
resultado 

resultado.isna().sum()

resultado[resultado['address']== '']

Se cuenta el número de sucursales que tiene cada negocio

y=metadata_2.groupby(by=['category']).count()
y

+ Se exporta a csv

resultado.to_csv('df_metadata_limpia.csv', index=False)
