==> 3_Análisis de sentimientos_california.ipynb <==
import pandas as pd 
from textblob import TextBlob

df_ny= pd.read_csv('df_california_limpio.csv')
df_ny

# Crear una nueva columna 'Sentiment' para almacenar los resultados del análisis de sentimiento
df_ny['Sentimiento'] = df_ny['Review'].apply(lambda text: TextBlob(str(text)).sentiment.polarity)


# Clasificar el sentimiento en categorías (positivo, neutral, negativo)
def categorizar_sentimiento(polarity):
    if polarity > 0:
        return 'Positivo'
    elif polarity == 0:
        return 'Neutral'
    else:
        return 'Negativo'

df_ny['Categ_Sentimiento'] = df_ny['Sentimiento'].apply(categorizar_sentimiento)
df_ny

df_ny.to_csv('df_california_analisis_sentimietnos.csv', index=False)
