==> 4_Merge.ipynb <==
import pandas as pd
import re 
import nltk
from nltk.corpus import stopwords
nltk.download('punkt')

# Lectura de archivos

df_california= pd.read_csv('df_california_analisis_sentimietnos.csv')
df_ny= pd.read_csv('df_ny_analisis_sentimietnos.csv')
df_georgia= pd.read_csv('df_georgia_analisis_sentimientos.csv')
df_colorado= pd.read_csv('df_colorado_analisis_sentimientos.csv')
df_texas= pd.read_csv('df_texas_analisis_sentimientos.csv')



metadata=pd.read_csv('metadata2.csv')
metadata

# Merge estado-metadata

merged_california = df_california.merge(metadata, on='gmap_id')
merged_california

merged_ny = df_ny.merge(metadata, on='gmap_id')
merged_ny

merged_georgia = df_georgia.merge(metadata, on='gmap_id')
merged_georgia

merged_colorado = df_colorado.merge(metadata, on='gmap_id')
merged_colorado

merged_texas = df_texas.merge(metadata, on='gmap_id')
merged_texas

# Checamos nuevamente si existen nan y se eliminan

merged_california['Review'].isnull().sum()

merged_california = merged_california.dropna(subset=['Review'])

merged_ny['Review'].isnull().sum()

merged_ny=merged_ny.dropna(subset=['Review'])

merged_texas['Review'].isnull().sum()

merged_texas=merged_texas.dropna(subset=['Review'])

merged_colorado['Review'].isnull().sum()

merged_colorado=merged_colorado.dropna(subset=['Review'])

merged_georgia['Review'].isnull().sum()

merged_georgia=merged_georgia.dropna(subset=['Review'])

# Convertir a string la columna Review

merged_california['Review'] = merged_california['Review'].astype(str)
merged_ny['Review'] = merged_ny['Review'].astype(str)
merged_texas['Review'] = merged_texas['Review'].astype(str)
merged_colorado['Review'] = merged_colorado['Review'].astype(str)
merged_georgia['Review'] = merged_georgia['Review'].astype(str)


merged_california['Review'] = merged_california['Review'].apply(lambda x: re.sub(r'[^\w\s]', '', x))
merged_ny['Review'] = merged_ny['Review'].apply(lambda x: re.sub(r'[^\w\s]', '', x))
merged_texas['Review'] = merged_texas['Review'].apply(lambda x: re.sub(r'[^\w\s]', '', x))
merged_colorado['Review'] = merged_colorado['Review'].apply(lambda x: re.sub(r'[^\w\s]', '', x))
merged_georgia['Review'] = merged_georgia['Review'].apply(lambda x: re.sub(r'[^\w\s]', '', x))


# Se renombran las columnas 

nombres={'Usuario': 'Usuario',
                                         'Rating_Usuario':'Rating_Usuario',
                                         'Review':'Review',
                                         'gmap_id':'gmap_id',
                                         'Anio':'Anio',
                                         'Mes':'Mes',
                                         'dia':'Dia',
                                         'Estado':'Estado',
                                         'Sentimiento':'Sentimiento',
                                         'Categ_Sentimiento':'Categ_Sentimiento',
                                         'name':'Negocio',
                                         'address':'Direccion_Negocio',
                                         'avg_rating':'Calificacion_promedio',
                                         'category':'Categoria',
                                         'num_of_reviews':'Numero_reviews',
                                         'hours':'Horario',
                                         'filtrado':'Filtrado',
                                         'clasificacion':'Clasificacion Negocio'
                                         }

merged_california= merged_california.rename(columns=nombres)
merged_ny= merged_ny.rename(columns=nombres)
merged_texas= merged_texas.rename(columns=nombres)
merged_colorado= merged_colorado.rename(columns=nombres)
merged_georgia= merged_georgia.rename(columns=nombres)


# Se exportan a csv

merged_california.to_csv('California.csv',index=False)
merged_ny.to_csv('New_York.csv',index=False)
merged_texas.to_csv('Texas.csv',index=False)
merged_colorado.to_csv('Colorado.csv',index=False)
merged_georgia.to_csv('Georgia.csv',index=False)
