==> 5_Clasificacion_Comentarios.ipynb <==
import pandas as pd
import nltk
from nltk.corpus import stopwords
nltk.download('punkt')

Lectura de archivos

df_california= pd.read_csv('California.csv')
df_ny= pd.read_csv('New_York.csv')
df_georgia= pd.read_csv('Georgia.csv')
df_colorado= pd.read_csv('Colorado.csv')
df_texas= pd.read_csv('Texas.csv')

df_california

# Quitamos las stopwords para la columna review

import nltk
nltk.download('stopwords')
stopwords = nltk.corpus.stopwords.words(['spanish', 'english'])

# Función para tokenizar y filtrar stopwords
def tokenize_and_filter(text):
    words = text.split()  # Divide la cadena en palabras
    filtered_words = [word for word in words if word.lower() not in stopwords]  # Filtra las stopwords
    return filtered_words

import re
# df_california['Review_2'] = df_california['Review'].apply(lambda x: re.sub(r'[^\w\s]', '', x))
df_california['filtrado'] = df_california['Review'].apply(tokenize_and_filter)

Que nos muestre las 300 palabras que más se repiten 

top_N=300
words=df_california.filtrado.tolist()
words=[word for list_ in words for word in list_]
words

word_dist= nltk.FreqDist(words)
rslt=pd.DataFrame(word_dist.most_common(top_N),columns=['Word','Frecuencia'])
rslt

rslt[251:300]

categorias_palabras_clave = {
    'Limpieza': ['clean', 'tidy', 'neat', 'hygienic'],
    'Comida': ['food', 'cuisine', 'menu', 'dishes','delicious','fresh','eat','drinks','tacos','menu','tasty','taste','cake'],
    'Tiempo': ['time', 'wait', 'quick','open','line','appointment'],
    'Precio': ['price', 'cost', 'expensive', 'affordable','money','pay','cheap'],
    'servicio':['helpful','employees','rude','service'],
    'Estacionamiento':['park'],
    'Mascota':['dog'],
    'Lugar':['lake'],

}

def clasificar_categorias(categorias):
    for categoria, palabras_clave in categorias_palabras_clave.items():
        for palabra_clave in palabras_clave:
            if palabra_clave in categorias:
                return categoria
    return 'Otros'  # Si no se encuentra ninguna categoría, se clasifica como 'otros'

df_california

df_california['Clasificacion Review'] = df_california['filtrado'].apply(clasificar_categorias)

df_california2=df_california[['Review','Categ_Sentimiento','Categoria','Clasificacion Negocio','Clasificacion Review']]
df_california2

df_california[df_california['Clasificacion Negocio']== 'Restaurante']

df_california.columns

df_california2=df_california[['Usuario', 'Rating_Usuario', 'Review', 'gmap_id', 'Anio', 'Mes', 'Dia',
       'Estado', 'Sentimiento', 'Categ_Sentimiento', 'Negocio',
       'Direccion_Negocio', 'Calificacion_promedio', 
       'Numero_reviews', 'Horario', 'Filtrado', 'Clasificacion Negocio', 'Clasificacion Review']]
df_california2

df_california2.to_csv('df_california_clasificacion_comentarios.csv',index= False )
