==> 6_Merge.ipynb <==
import pandas as pd

df_california=pd.read_csv('df_california_clasificacion_comentarios.csv')
df_ny=pd.read_csv('df_ny_clasificacion_comentarios.csv')
df_texas=pd.read_csv('df_texas_clasificacion_comentarios.csv')
df_colorado=pd.read_csv('df_colorado_clasificacion_comentarios.csv')
df_georgia=pd.read_csv('df_georgia_clasificacion_comentarios.csv')


# Merge de todos los dataframes

df_google_maps = pd.concat([df_california, df_ny,df_texas,df_colorado,df_georgia])
df_google_maps


df_google_maps.reset_index(drop=True, inplace=True)
df_google_maps

['kfc','mcdonalds','subway','pizza hut','dominos pizza']

df_google_maps.to_csv('df_google_maps.csv',index=False)
